refactor: 구조 개선 - 컴포넌트 & 페이지 관리 명확하게 수정(sign & findUser)
