--- src/components/Sign/index.ts
@@ -1,6 +1,0 @@
-import SignIn from './Sign-in';
-import FindIdByPhone from './Find/FindIdByPhone';
-import FindIdByEmail from './Find/FindIdByEmail';
-
-
-export { SignIn, FindIdByPhone, FindIdByEmail };
