+++ src/pages/FindUser/id/index.ts
@@ -1,0 +1,4 @@
+import FindIdByEmail from './FindIdByEmail';
+import FindIdByPhone from './FindIdByPhone';
+
+export { FindIdByEmail, FindIdByPhone };
