+++ src/pages/FindUser/index.ts
@@ -1,0 +1,3 @@
+import FindUserId from './FindUserId';
+
+export { FindUserId };
