+++ src/pages/Sign/index.ts
@@ -1,0 +1,4 @@
+import SignInPage from './SignInPage';
+import SignUpPage from './SignUpPage';
+
+export { SignInPage, SignUpPage };
