+++ src/pages/index.ts
@@ -3,8 +3,9 @@ import ErrorPage from './error-page';
 import MainPage from './main';
 
 /* Sign */
-import SignInPage from './Sign/SignInPage';
-import SignUpPage from './Sign/SignUpPage';
-import FindUserId from './Sign/FindUserId';
+import { SignInPage, SignUpPage } from './Sign';
+
+/* Find User */
+import { FindUserId } from './FindUser';
 
 export { ErrorPage, SignInPage, SignUpPage, FindUserId, MainPage };
